First commit : -Created Social component to hold Twitter and Discord logos. Design can be passed as a parameter with <Social design={} /> -Created Header component to hold all content which will be on top of every page -Updated Footer to use new Social component, renamed AccordionDemo to AccordionDonation for clarity -Grouped all Accordion designs in one file (Accordion.tsx). Just import the one you want from the file -Removed AccordionHeader.tsx, content is in Accordion.tsx now -Updated main and blockchain page to use new components -Added .env to the gitignore file -Removed unused imports
