+++ components/Header.tsx
@@ -1,0 +1,17 @@
+import { Social } from "./Social"
+
+export function Header({ design }: { design: string }) {
+    return (
+        <div className={design}>
+            <p className="font-garamond text-gray-900 font-serif text-3xl font-medium normal-case not-italic no-underline leading-tight tracking-tighter">
+                Encyclochain
+            </p>
+            <p className=" font-garamond text-gray-900 font-serif text-3xl w-full text-center font-medium normal-case not-italic no-underline leading-tight tracking-tighter max-lg:hidden">
+                Blockchains encyclopedia
+            </p>
+            <p>
+                <Social design="flex gap-[15px] max-md:w-full max-md:flex max-md:justify-center" />
+            </p>
+        </div>
+    )
+}
